requirements.txt
google-genai==1.18.0
google-generativeai==0.8.3
flask==2.3.2
langchain-community==0.3.7
typer==0.12.3
chromadb==0.6.3
uvicorn[standard]